requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-documentai
google-cloud-firestore
google-cloud-storage
numpy
pandas
pdfkit
Flask
Flask-SQLAlchemy
Flask-WTF
gunicorn
pyopenssl
pyarrow
